% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.r
\name{decode_car}
\alias{decode_car}
\title{Decode a Content Addressable aRchive (CAR) file}
\usage{
decode_car(data)
}
\arguments{
\item{data}{A raw vector containing a CAR file}
}
\value{
A list with header information and decoded blocks
}
\description{
This function decodes a CAR file from a raw vector, extracting the header and blocks.
}
\examples{
car_file <- system.file("extdata", "sample.car", package = "libipldr")
car_data <- readBin(car_file, what = "raw", n = file.size(car_file))
decode_car(car_data)
}
