% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/r-wrappers.r
\name{decode_cid}
\alias{decode_cid}
\title{Decode a CID string to its components}
\usage{
decode_cid(cid_str)

decode_cid(cid_str)
}
\arguments{
\item{cid_str}{A string containing a valid CID}
}
\value{
A list with CID components
}
\description{
This function decodes a CID string into its components (version, codec, and hash).
}
\examples{
\dontrun{
# Decode a CID:
cid_info <- decode_cid("bafyreib775pirw4o3rz4iwdjwi3rz7q4z5t4xjyfrwnk2yukhzo2wyr4ye")
}
}
