% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.r
\name{decode_dag_cbor}
\alias{decode_dag_cbor}
\title{Decode DAG-CBOR encoded data to an R object}
\usage{
decode_dag_cbor(data)
}
\arguments{
\item{data}{A raw vector containing DAG-CBOR encoded data}
}
\value{
An R object representing the decoded data
}
\description{
This function decodes a raw vector containing DAG-CBOR encoded data into an R
object. DAG-CBOR is a deterministic subset of the CBOR format, used by IPFS
and AtProto (Bluesky) for data representation.
}
\examples{
# Decode a simple DAG-CBOR map {"a": "Hello", "b": "World!"}
cbor_data <- as.raw(c(
  0xa2, 0x61, 0x61, 0x65, 0x48, 0x65, 0x6c, 0x6c, 0x6f,
  0x61, 0x62, 0x66, 0x57, 0x6f, 0x72, 0x6c, 0x64, 0x21
))
decode_dag_cbor(cbor_data)
}
