% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{libipldr-package}
\alias{libipldr}
\alias{libipldr-package}
\title{libipldr: R Bindings to the 'Rust' 'IPLD' Library}
\description{
Provides R bindings to decode DAG-CBOR (Directed Acyclic Graph Concise Binary Object Representation) encoded data, CIDs (Content Identifiers), and CAR (Content Addressable aRchive) files using the 'Rust' 'IPLD' (InterPlanetary Linked Data) library \url{https://github.com/ipld/libipld}. This is especially useful for working with data from 'IPFS' (InterPlanetary File System) and 'AtProto' (Bluesky) applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JBGruber/libipldr}
  \item Report bugs at \url{https://github.com/JBGruber/libipldr/issues}
}

}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})

Authors:
\itemize{
  \item Johannes B. Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID})
}

Other contributors:
\itemize{
  \item Ilya Siamionau (Author of the original python-libipld Rust implementation) [contributor]
}

}
