% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.r
\name{print.car_file}
\alias{print.car_file}
\title{Print method for CAR file objects}
\usage{
\method{print}{car_file}(x, ...)
}
\arguments{
\item{x}{A CAR file object from decode_car()}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for CAR file objects
}
