% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-wrappers.r
\name{print.cid}
\alias{print.cid}
\title{Print method for CID objects}
\usage{
\method{print}{cid}(x, ...)
}
\arguments{
\item{x}{A CID object from decode_cid()}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for CID objects
}
