% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/r-wrappers.r
\name{decode_car}
\alias{decode_car}
\title{Decode a CAR file byte vector}
\usage{
decode_car(data)

decode_car(data)
}
\arguments{
\item{data}{A raw vector containing a CAR file}
}
\value{
A list with header information and decoded blocks
}
\description{
This function decodes a CAR file from a raw vector, extracting the header and blocks.
}
\examples{
\dontrun{
# When you have a CAR file as raw data:
car_data <- decode_car(raw_car_data)
}
}
