% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/r-wrappers.r
\name{decode_dag_cbor}
\alias{decode_dag_cbor}
\title{Decode a DAG-CBOR encoded byte vector to an R object}
\usage{
decode_dag_cbor(data)

decode_dag_cbor(data)
}
\arguments{
\item{data}{A raw vector containing DAG-CBOR encoded data}
}
\value{
An R object representing the decoded data
}
\description{
This function decodes a raw vector containing DAG-CBOR encoded data into an R
object. DAG-CBOR is a deterministic subset of the CBOR format, used by IPFS
and AtProto (Bluesky) for data representation.
}
\examples{
\dontrun{
# When you have DAG-CBOR encoded data:
result <- decode_dag_cbor(raw_data)
}
}
