% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/r-wrappers.r
\name{decode_dag_cbor_multi}
\alias{decode_dag_cbor_multi}
\title{Decode multiple DAG-CBOR objects from a byte vector}
\usage{
decode_dag_cbor_multi(data)

decode_dag_cbor_multi(data)
}
\arguments{
\item{data}{A raw vector containing multiple DAG-CBOR encoded objects}
}
\value{
A list of R objects, each representing a decoded DAG-CBOR object
}
\description{
This function decodes multiple consecutive DAG-CBOR objects from a single raw vector.
}
\examples{
\dontrun{
# When you have multiple DAG-CBOR objects in one stream:
results <- decode_dag_cbor_multi(raw_data)
}
}
