% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{libipldr-package}
\alias{libipldr}
\alias{libipldr-package}
\title{libipldr: R Bindings to the Rust IPLD Library}
\description{
Provides R bindings to decode DAG-CBOR encoded data, CIDs (Content Identifiers), and CAR (Content Addressable aRchive) files using Rust's IPLD library. This is especially useful for working with data from IPFS and AtProto (Bluesky) applications.
}
\author{
\strong{Maintainer}: Johannes B. Gruber \email{JohannesB.Gruber@gmail.com}

}
